% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosetta.R
\name{best_designs}
\alias{best_designs}
\title{Return a list of the top designs in a Rosetta scorefile, based on a scoreterm of choice.}
\usage{
best_designs(datatable, scoreterm = "total_score", top = 1, number = 50,
  descending = FALSE)
}
\arguments{
\item{datatable}{A table containing a Rosetta-style scorefile data.}

\item{scoreterm}{The name of the scoreterm to be used to sort.}

\item{top}{The first design to include.  For example, if it is 3, the top two designs will be skipped.}

\item{number}{The total number of designs to list.}

\item{descending}{If true, the highest score will be considered the "best."  If false (default), the lowest score will be considered the "best."}
}
\value{
A list of the design names requested.
}
\description{
Return a list of the top designs in a Rosetta scorefile, based on a scoreterm of choice.
}
\details{
The first and last design can be customized.
}
\seealso{
\code{\link{score_plot}} \code{\link{read_rosetta_score}}
}
