% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_functions.R
\name{mm_kinetics}
\alias{mm_kinetics}
\title{This is a master function that does complete Michaelis-Menten kinetics analysis.}
\usage{
mm_kinetics(raw_data, infile = "", outfile = "")
}
\arguments{
\item{raw_data}{A data table containing time data in the leftmost column and activity data in the other columns.  The top row should be the concentration series.  Row and column names are not used (i.e. time data should not be the row names, and concentration data should not be the column names).}

\item{infile}{An optional filename of a CSV file containing the results of a previous mm_kinetics run.}

\item{outfile}{An optional filename of a CSV file that will be written with the results of this mm_kinetics run, for future re-analysis.}
}
\value{
The nls object generated in the final round of non-linear regression.
}
\description{
The raw data kinetic data, as a single table, is passed to the function.  Each data series is analyzed in sequence, allowing the user to select appropriate minimum and maximum limits for calculating the initial rate.
Once all data series are analyzed, the Michaelis-Menten curve (rate vs. substrate concentration) is plotted and non-linear regression is performed.
The user will then be prompted to reanalyze any of the kinetic series, or to exclude one or more series from analysis.
Once the user is satisfied with the results, the function ends, printing and returning the non-linear regression object.
If the optional \code{infile} string is given, a CSV file containing a previous set of results for \code{raw_data} is loaded.
In this case, the initial analysis of the data series is skipped.
If the optional \code{outfile} string is given, a CSV file is written containing the final results of the analysis.
\code{infile} needs to be generated by \code{mm_kinetics} to be a valid result file.
}
