% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_functions.R
\name{mm_nls}
\alias{mm_nls}
\title{Performs a non-linear regression fit to the Michaelis-Menten equation and prints and graphs the results.}
\usage{
mm_nls(concentration, rates, exclude = c(FALSE))
}
\arguments{
\item{concentration}{The substrate concentrations used in the non-linear regression.}

\item{rates}{The rates corresponding to the substrate concentrations given in \code{concentration}.}

\item{exclude}{An optional flag indicating that this datapoint should be excluded from the non-linear regression analysis.}
}
\value{
The non-linear regression object is returned.
}
\description{
Performs a non-linear regression fit to the Michaelis-Menten equation and prints and graphs the results.
}
\seealso{
\code{\link{nls}}
}
