% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_rosetta_score}
\alias{read_rosetta_score}
\title{Loads a Rosetta-style score file into a data frame.}
\usage{
read_rosetta_score(scorefile)
}
\arguments{
\item{scorefile}{The path and filename of the Rosetta-style scorefile to load.}
}
\value{
A data frame object containing the data from the scorefile given in \code{scorefile}.
}
\description{
Loads a Rosetta-style score file into a data frame.
}
\seealso{
\code{\link{read.table}}
}
