% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_tecan}
\alias{read_tecan}
\title{Loads the contents of Tecan-generated Excel sheets into a data frame.}
\usage{
read_tecan(excelfile, excelsheet)
}
\arguments{
\item{excelfile}{The name of the Excel file containing the sheet.}

\item{excelsheet}{The name of the sheet in the Excel file containing your data.}
}
\value{
A data frame containing the Tecan data.
}
\description{
Loads the contents of Tecan-generated Excel sheets into a data frame.
}
\seealso{
\code{\link[xlsx]{read.xlsx}}
}
