% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_functions.R
\name{regression_loop}
\alias{regression_loop}
\title{Calculate the slope of a single data series.}
\usage{
regression_loop(kineticdata, timeheader, activityheader,
  startreg = min(as.numeric(as.character(kineticdata[-1, timeheader]))),
  endreg = max(as.numeric(as.character(kineticdata[-1, timeheader]))))
}
\arguments{
\item{kineticdata}{A dataframe that contains one or more kinetic series.  The column names of the dataframe should be the name of each dataset.}

\item{timeheader}{A string containing the column name of the time data for the kinetic series.}

\item{activityheader}{A string containing the activity (or concentration) data for the kinetic series.}

\item{startreg}{A numeric value indicating the smallest timepoint to be used in the first round of regression.  If omitted, the smallest timepoint is used.  If the timepoint is not present in the dataset, the smallest timepoint greater than \code{startreg} is used.}

\item{endreg}{A numeric value indicating the largest timepoint to be used in the first round of regression.  If omitted, the largest timepoint is used.  If the timepoint is not present in the dataset, the largest timepoint smaller than \code{endreg} is used.}
}
\value{
A list object containing, from the final round of linear regression, (1) the slope, (2) the intercept, (3) the lower timepoint limit, (4) the higher timepoint limit, and (5) the complete regression object.
}
\description{
This function calculates the slope for a single kinetic data series.
The user will then be prompted to enter alternate x-limits for the dataset.  The data will be re-plotted with the new regression curve.
If the user hits enter for either the minimum or maximum limit, the previous value will be used.
When the user hits enter for both the minimum and maximum limit, the function will end.
}
\seealso{
\code{\link{lm}}
}
