% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosetta.R
\name{ros_plot}
\alias{ros_plot}
\title{Plots a graph of two score terms from a score file.}
\usage{
ros_plot(datatable, xterm = "rmsd", yterm = "total_score", title = "")
}
\arguments{
\item{datatable}{A table containing a Rosetta-style scorefile data.}

\item{xterm}{The name of the scoreterm to be plotted on the x-axis.  Defaults to "rmsd".}

\item{yterm}{The name of the scoreterm to be plotted on the y-axis.  Defaults to "total_score".}

\item{title}{A title for the plot, if desired.}
}
\description{
Plots a graph of two score terms from a score file.
}
\details{
If the axes are not given, assume they are total_score vs. rmsd.
}
\seealso{
\code{\link{score_plot}}
}
