% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosetta.R
\name{score_plot}
\alias{score_plot}
\title{Loads a Rosetta-style scorefile (*.sc) and generates a plot.}
\usage{
score_plot(scorefile, xterm = "rmsd", yterm = "total_score",
  title = scorefile)
}
\arguments{
\item{scorefile}{The path and filename of the Rosetta-style scorefile to load.}

\item{xterm}{The name of the scoreterm to be plotted on the x-axis.  Defaults to "rmsd".}

\item{yterm}{The name of the scoreterm to be plotted on the y-axis.  Defaults to "total_score".}

\item{title}{A title for the plot, if desired.}
}
\value{
A data frame object containing the data from the scorefile given in \code{scorefile}.
}
\description{
Loads a Rosetta-style scorefile (*.sc) and generates a plot.
}
\details{
If the axes are not given, assume they are total_score vs. rmsd.
}
\seealso{
\code{\link{ros_plot}}, \code{\link{read_rosetta_score}}
}
